requirements.txt
pyautogui==0.9.54
pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
pytesseract==0.3.10
# Windows API支持
pywin32==306